refactor(check): the judgeable-argument seam moves to the analyzer
The documentation-host advisory was the line that crossed nika-check's
15k wall (15032 · vector 24) and the credited-gate test the one that
crossed lib.rs's 1500 (1539 · vector 12). The descent, not a trim:
`ConstStrings` · `judgeable_arg` · `url_host` · `templated_url_host`
move verbatim into `nika-check-analyzer::static_args` (the same carve as
`silent_literal`: the seam reads the AST alone), re-exported at their old
paths so every consumer keeps its spelling; the test moves to the hints
test module next to the hint it proves.

nika-check 15032 → 14912 prod LOC · lib.rs 1539 → 1491 lines ·
nika-check-analyzer 3791 → 3932 · its public-api baseline carries the
new module (no Linux-only lines in that crate) · nika-check 568 ·
nika-check-analyzer 180 lib tests · clippy --all-targets -D warnings
clean.
